requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scipy
opencv-python
